added p_excess calculation to daily calculations and set initial p_animal
