+++ main.py
@@ -10,6 +10,8 @@ Author(s): Kass Chupongstimun, kass_c@hotmail.com
 # !/usr/bin/env python3
 
 import RUFAS
+import numpy
+import random
 
 def main():
 
@@ -43,4 +45,6 @@ def main():
 # PROGRAM ENTRY POINT
 # -------------------------------------------------------------------------------
 if __name__ == '__main__':
+    numpy.random.seed(9)
+    random.seed(9)
     main()
